+++ docs/versionhistory.rst
@@ -10,6 +10,7 @@ APScheduler, see the :doc:`migration section <migration>`.
   (`#1059 <https://github.com/agronholm/apscheduler/issues/1059>`_)
 
 **4.0.0a6**
+
 - **BREAKING** Refactored ``AsyncpgEventBroker`` to directly accept a connection string,
   thus eliminating the need for the ``AsyncpgEventBroker.from_dsn()`` class method
 - **BREAKING** Added the ``extend_acquired_schedule_leases()`` data store method to